CU-37t2pvw: admin specific model for easy export uploading
