+++ webapp/api/api/forms.py
@@ -1,10 +1,12 @@
+import json
 import os
 
+from django.db.models.fields.files import FileField
 from django.db.models.signals import post_save, post_delete, pre_save
 from django.dispatch import receiver
 
-from api.data_utils import dataset_from_file, delete_orphan_docs
-from api.models import Dataset
+from api.data_utils import dataset_from_file, delete_orphan_docs, upload_projects_export
+from api.models import Dataset, ExportedProject
 
 
 @receiver(post_save, sender=Dataset)
@@ -27,3 +29,10 @@ def remove_dataset_file(sender, instance, **kwargs):
     if instance.original_file:
         if os.path.isfile(instance.original_file.path):
             os.remove(instance.original_file.path)
+
+
+@receiver(post_save, sender=ExportedProject)
+def save_exported_projects(sender, instance, **kwargs):
+    if not instance.trainer_export_file.path.endswith('.json'):
+        raise Exception("Please make sure the file is a .json file")
+    upload_projects_export(json.load(open(instance.trainer_export_file.path)))
